// web/src/store/schema.gen.ts
export type ProcessesRecord = {
  id: string
  order: string
  baseline_staff: unknown
  code: string
  demand: unknown
  durations: unknown
  is_baseline: boolean
  name: string
  point_ids: string[]
  sla: unknown
  task_type: 'pallet_inbound' | 'pallet_putaway' | 'pallet_outbound' | 'piece_pick' | 'pallet_move' | 'cleaning'
}